requirements.txt
pytesseract==0.3.13
pdfplumber==0.11.4
python-docx==1.1.2
googletrans==4.0.0-rc1
flask==3.0.3
Pillow==10.4.0
streamlit==1.39.0